requirements.txt
streamlit 
langchain-groq 
langchain-core
langchain-community
chromadb
pypdf
unstructured
python-docx
tavily-python
python-dotenv